# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BT资产化相关API",
        "title": "BT CLI",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8001",
    "basePath": "/",
    "paths": {
        "/v1/buildolotx": {
            "post": {
                "description": "生成OLO转账交易",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "生成OLO转账交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.BuildOLOTxReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BuildTxResult"
                        }
                    }
                }
            }
        },
        "/v1/buildtokenbatchtx": {
            "post": {
                "description": "生成代币批量转账交易，转账金额相同",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "生成代币批量转账交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.BuildTokenBatchTransferTxReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BuildTxResult"
                        }
                    }
                }
            }
        },
        "/v1/buildtokenbatchtxs": {
            "post": {
                "description": "生成代币批量转账交易，转账金额不同",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "生成代币批量转账交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.BuildTokenBatchTransfersTxReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BuildTxResult"
                        }
                    }
                }
            }
        },
        "/v1/buildtokentx": {
            "post": {
                "description": "生成代币转账交易",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "生成代币转账交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.BuildTokenTxReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BuildTxResult"
                        }
                    }
                }
            }
        },
        "/v1/checktx": {
            "get": {
                "description": "检测交易是否成功上链",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "检测交易是否成功上链",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易hash",
                        "name": "hash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/v1/genkey": {
            "get": {
                "description": "生成mondo账户，该账户默认未上链，需要通过转账交易激活",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "生成账户",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.GenKeyResult"
                        }
                    }
                }
            }
        },
        "/v1/olobalance": {
            "get": {
                "description": "查询地址OLO余额，如果地址不存在，返回余额为0",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "查询地址OLO余额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BalanceResult"
                        }
                    }
                }
            }
        },
        "/v1/sendtx": {
            "post": {
                "description": "发送交易",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "发送交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SignedEvmTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/v1/tokenbalance": {
            "get": {
                "description": "查询地址代币余额",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "查询地址代币余额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "代币合约地址",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.BalanceResult"
                        }
                    }
                }
            }
        },
        "/v1/validaddress": {
            "get": {
                "description": "校验地址是否合法",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "校验地址是否合法",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.BalanceResult": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                }
            }
        },
        "main.BuildOLOTxReq": {
            "type": "object",
            "properties": {
                "priKey": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.BuildTokenBatchTransferTxReq": {
            "type": "object",
            "properties": {
                "priKey": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "tos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.BuildTokenBatchTransfersTxReq": {
            "type": "object",
            "properties": {
                "priKey": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "tos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.To"
                    }
                }
            }
        },
        "main.BuildTokenTxReq": {
            "type": "object",
            "properties": {
                "priKey": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.BuildTxResult": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "signedEvmTx": {
                    "type": "string"
                }
            }
        },
        "main.GenKeyResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "priKey": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                }
            }
        },
        "main.SignedEvmTx": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "load": {
                            "description": "hex编码",
                            "type": "string"
                        },
                        "memo": {
                            "description": "备注",
                            "type": "string"
                        },
                        "to": {
                            "description": "合约地址",
                            "type": "string"
                        },
                        "value": {
                            "type": "string"
                        }
                    }
                },
                "createdAt": {
                    "description": "时间戳unixNano",
                    "type": "integer"
                },
                "gasLimit": {
                    "type": "integer"
                },
                "gasPrice": {
                    "type": "string"
                },
                "mode": {
                    "description": "模式:0-default/commit 1-async 2-sync",
                    "type": "integer"
                },
                "nonce": {
                    "type": "integer"
                },
                "sender": {
                    "description": "pubkey",
                    "type": "string"
                },
                "signature": {
                    "description": "hex编码",
                    "type": "string"
                }
            }
        },
        "main.To": {
            "type": "object",
            "properties": {
                "to": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}